+++ modules/coverage/geotk-coverageio-imagery/src/main/java/org/geotoolkit/image/io/large/LargeMap2D.java
@@ -23,6 +23,8 @@ import org.geotoolkit.image.io.XImageIO;
 import javax.imageio.ImageIO;
 import javax.imageio.ImageReader;
 import javax.imageio.ImageWriter;
+import javax.imageio.stream.ImageInputStream;
+import javax.imageio.stream.ImageOutputStream;
 import javax.media.jai.RasterFactory;
 import java.awt.*;
 import java.awt.image.*;
@@ -104,7 +106,6 @@ class LargeMap2D {
         //quad tree directory architecture.
         this.dirPath = TEMPORARY_PATH + "/img_"+ri.hashCode();
         this.qTD     = new QuadTreeDirectory(dirPath, numTilesX, numTilesY, FORMAT, true);
-        qTD.create4rchitecture();
 
         //reader writer
         this.imgReader = XImageIO.getReaderByFormatName(FORMAT, null, Boolean.FALSE, Boolean.TRUE);
@@ -197,17 +198,22 @@ class LargeMap2D {
             final File getFile = new File(qTD.getPath(tX, tY));
 
             if (getFile.exists()) {
-                imgReader.setInput(ImageIO.createImageInputStream(getFile));
-                final BufferedImage buff = imgReader.read(0);
-                imgReader.dispose();
-                //add in cache list.
-                final Raster wr = checkRaster(buff.getRaster(), tX, tY);
-                final long rastWeight = getRasterWeight(wr);
-                row.put(tX, new LargeRaster(tX, tY, rastWeight, wr));
-                stack.addLast(new Point(tX, tY));
-                remainingCapacity -= rastWeight;
-                checkMap();
-                result = wr;
+                final ImageInputStream stream = ImageIO.createImageInputStream(getFile);
+                if (stream != null) {
+                    imgReader.setInput(stream);
+                    final BufferedImage buff = imgReader.read(0);
+                    imgReader.setInput(null);
+                    imgReader.dispose();
+                    stream.close();
+                    //add in cache list.
+                    final Raster wr = checkRaster(buff.getRaster(), tX, tY);
+                    final long rastWeight = getRasterWeight(wr);
+                    row.put(tX, new LargeRaster(tX, tY, rastWeight, wr));
+                    stack.addLast(new Point(tX, tY));
+                    remainingCapacity -= rastWeight;
+                    checkMap();
+                    result = wr;
+                }
             }
         }
         return result;
@@ -281,9 +287,15 @@ class LargeMap2D {
     private void writeRaster(File path, Raster raster) throws IOException {
         final WritableRaster wr  = RasterFactory.createWritableRaster(raster.getSampleModel(), raster.getDataBuffer(), WPOINT);
         final BufferedImage rast = new BufferedImage(cm, wr, true, null);
-        imgWriter.setOutput(ImageIO.createImageOutputStream(path));
+        final ImageOutputStream stream = ImageIO.createImageOutputStream(path);
+        if (stream == null) {
+            throw new IOException("No output connexion can be opened to write tile.");
+        }
+        imgWriter.setOutput(stream);
         imgWriter.write(rast);
+        imgWriter.setOutput(null);
         imgWriter.dispose();
+        stream.close();
         path.deleteOnExit();
     }
